backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
drf-spectacular>=0.27
Pillow>=10.0
python-decouple>=3.8
